HPA gene query

Queries one if the HPA data sets with the id Ensembl gene identifier. The data set to be used is defined by the hpadata argument.

getHpa(id, hpadata = NULL, type = c("data", "details"))

Arguments

id

A Ensembl gene identifier.

hpadata

A character with the data set to query. One of available hpar datasets, available by calling allHparData (or any unambiguous prefix), or NULL (default). In the latter case, the default option is used. See setHparOptions and getHparOptions for more details.

type

A character defining what type data to return. One of data (default, for the data as data.frame) or details (to open the id's HPA webpage).

Value

A data.frame with the information corresponding to the respective id genes. If type is details, then the dataframe is returned invisibly and a web page is opened with browseURL.

Examples

id <- "ENSG00000000003"
## Define 'hpadata' data manually
getHpa(id, hpadata = "hpaSubcellularLoc")

#>              Gene Gene.name Main.location Other.location Expression.type
#> 1 ENSG00000000003    TSPAN6     Cytoplasm                            APE
#>   Reliability Main.location.GO.id Other.location.GO.id
#> 1   Uncertain          GO:0005737                     

head(getHpa(id, hpadata = "hpaNormalTissue"), )

#>              Gene Gene.name        Tissue           Cell.type        Level
#> 1 ENSG00000000003    TSPAN6 adrenal gland     glandular cells Not detected
#> 2 ENSG00000000003    TSPAN6      appendix     glandular cells       Medium
#> 3 ENSG00000000003    TSPAN6      appendix     lymphoid tissue Not detected
#> 4 ENSG00000000003    TSPAN6   bone marrow hematopoietic cells Not detected
#> 5 ENSG00000000003    TSPAN6        breast          adipocytes Not detected
#> 6 ENSG00000000003    TSPAN6        breast     glandular cells         High
#>   Expression.type Reliability
#> 1             APE   Uncertain
#> 2             APE   Uncertain
#> 3             APE   Uncertain
#> 4             APE   Uncertain
#> 5             APE   Uncertain
#> 6             APE   Uncertain

head(getHpa(id, hpadata = "rnaGeneTissue"))

#>              Gene Gene.name          Sample Value Unit    Abundance
#> 1 ENSG00000000003    TSPAN6  adipose tissue  20.5 FPKM       Medium
#> 2 ENSG00000000003    TSPAN6   adrenal gland  17.6 FPKM       Medium
#> 3 ENSG00000000003    TSPAN6        appendix   6.4 FPKM          Low
#> 4 ENSG00000000003    TSPAN6     bone marrow   0.4 FPKM Not detected
#> 5 ENSG00000000003    TSPAN6 cerebral cortex   9.3 FPKM          Low
#> 6 ENSG00000000003    TSPAN6           colon  41.2 FPKM       Medium

head(getHpa(id, hpadata = "rnaGeneCellLine"))

#>              Gene Gene.name  Sample Value Unit Abundance
#> 1 ENSG00000000003    TSPAN6   A-431  17.5 FPKM       Low
#> 2 ENSG00000000003    TSPAN6    A549  27.2 FPKM    Medium
#> 3 ENSG00000000003    TSPAN6  AN3-CA  32.9 FPKM    Medium
#> 4 ENSG00000000003    TSPAN6    BEWO  26.0 FPKM    Medium
#> 5 ENSG00000000003    TSPAN6  CACO-2  56.2 FPKM      High
#> 6 ENSG00000000003    TSPAN6 CAPAN-2  30.0 FPKM    Medium

head(getHpa(id, hpadata = "hpaCancer"))

#>              Gene Gene.name         Tumor        Level Count.patients
#> 1 ENSG00000000003    TSPAN6 breast cancer         High              1
#> 2 ENSG00000000003    TSPAN6 breast cancer       Medium              7
#> 3 ENSG00000000003    TSPAN6 breast cancer          Low              2
#> 4 ENSG00000000003    TSPAN6 breast cancer Not detected              2
#> 5 ENSG00000000003    TSPAN6     carcinoid         High              0
#> 6 ENSG00000000003    TSPAN6     carcinoid       Medium              1
#>   Total.patients Expression.type
#> 1             12        Staining
#> 2             12        Staining
#> 3             12        Staining
#> 4             12        Staining
#> 5              4        Staining
#> 6              4        Staining

## Sets default to "SubcellularLoc"
setHparOptions(hpadata = "hpaSubcellularLoc")
getHpa(id)       ## now uses "hpaSubcellularLoc"

#>              Gene Gene.name Main.location Other.location Expression.type
#> 1 ENSG00000000003    TSPAN6     Cytoplasm                            APE
#>   Reliability Main.location.GO.id Other.location.GO.id
#> 1   Uncertain          GO:0005737                     

setHparOptions() ## reset to "hpaNormalTissue"
head(getHpa(id))

#>              Gene Gene.name        Tissue           Cell.type        Level
#> 1 ENSG00000000003    TSPAN6 adrenal gland     glandular cells Not detected
#> 2 ENSG00000000003    TSPAN6      appendix     glandular cells       Medium
#> 3 ENSG00000000003    TSPAN6      appendix     lymphoid tissue Not detected
#> 4 ENSG00000000003    TSPAN6   bone marrow hematopoietic cells Not detected
#> 5 ENSG00000000003    TSPAN6        breast          adipocytes Not detected
#> 6 ENSG00000000003    TSPAN6        breast     glandular cells         High
#>   Expression.type Reliability
#> 1             APE   Uncertain
#> 2             APE   Uncertain
#> 3             APE   Uncertain
#> 4             APE   Uncertain
#> 5             APE   Uncertain
#> 6             APE   Uncertain

## multiple ids
getHpa(id = c("ENSG00000000003", "ENSG00000000005"),
       hpadata = "hpaSubcellularLoc")

#>              Gene Gene.name Main.location Other.location Expression.type
#> 1 ENSG00000000003    TSPAN6     Cytoplasm                            APE
#>   Reliability Main.location.GO.id Other.location.GO.id
#> 1   Uncertain          GO:0005737                     

## Not run: ------------------------------------
# ## opens a browser with http://www.proteinatlas.org/ENSG00000163435
# getHpa("ENSG00000163435", type = "details") 
## ---------------------------------------------