Obtain HPA data release information

Obtain the data and release information of the Human Protein Atlas (HPA) sets distributed by `hpar`. See http://www.proteinatlas.org/about/releases for the HPA release history.

getHpaDate()

Value

A `character()` with the release information.

Author

Laurent Gatto

Examples

allHparData()
#>                    Title
#> 1          hpaCancer16.1
#> 2              hpaCancer
#> 3    hpaNormalTissue16.1
#> 4        hpaNormalTissue
#> 5           hpaSecretome
#> 6    hpaSubcellularLoc14
#> 7  hpaSubcellularLoc16.1
#> 8      hpaSubcellularLoc
#> 9     rnaConsensusTissue
#> 10       rnaFantomTissue
#> 11   rnaGeneCellLine16.1
#> 12       rnaGeneCellLine
#> 13     rnaGeneTissue21.0
#> 14         rnaGtexTissue
#> 15          rnaHpaTissue
#>                                                                                                                                                  Description
#> 1                                                                       Pathology data: Staining profiles for proteins in human tumor tissue (version 16.1).
#> 2                                                                                      Pathology data: Staining profiles for proteins in human tumor tissue.
#> 3                Normal tissue data: Expression profiles for proteins in human tissues based on immunohistochemisty using tissue micro arrays (version 16.1)
#> 4                              Normal tissue data: Expression profiles for proteins in human tissues based on immunohistochemisty using tissue micro arrays.
#> 5                                                Secretome data: Ensembl genes with at least one predicted secreted transcript according to HPA predictions.
#> 6                                       Subcellular location data: Subcellular location of proteins based on immunofluorescently stained cells (version 14).
#> 7                                     Subcellular location data: Subcellular location of proteins based on immunofluorescently stained cells (version 16.1).
#> 8                                                                               Subcellular location of proteins based on immunofluorescently stained cells.
#> 9  RNA consensus tissue gene data: Consensus transcript expression levels summarized per gene in 54 tissues based on transcriptomics data from HPA and GTEx.
#> 10                                            RNA FANTOM tissue gene data: Transcript expression levels summarized per gene in 60 tissues based on CAGE data
#> 11                                            RNA HPA cell line gene data: Transcript expression levels summarized per gene in 69 cell lines (version 16.1).
#> 12                                                           RNA HPA cell line gene data: Transcript expression levels summarized per gene in 69 cell lines.
#> 13                                  RNA HPA tissue gene data: Transcript expression levels summarized per gene in 37 tissues based on RNA-seq (version 21.0)
#> 14                                                RNA GTEx tissue gene data: Transcript expression levels summarized per gene in 37 tissues based on RNA-seq
#> 15                                               RNA HPA tissue gene data: Transcript expression levels summarized per gene in 256 tissues based on RNA-seq.
#>    BiocVersion Genome SourceType                             SourceUrl
#> 1         3.16     NA        TSV https://www.proteinatlas.org/download
#> 2         3.16     NA        TSV https://www.proteinatlas.org/download
#> 3         3.16     NA        TSV https://www.proteinatlas.org/download
#> 4         3.16     NA        TSV https://www.proteinatlas.org/download
#> 5         3.16     NA        TSV https://www.proteinatlas.org/download
#> 6         3.16     NA        TSV https://www.proteinatlas.org/download
#> 7         3.16     NA        TSV https://www.proteinatlas.org/download
#> 8         3.16     NA        TSV https://www.proteinatlas.org/download
#> 9         3.16     NA        TSV https://www.proteinatlas.org/download
#> 10        3.16     NA        TSV https://www.proteinatlas.org/download
#> 11        3.16     NA        TSV https://www.proteinatlas.org/download
#> 12        3.16     NA        TSV https://www.proteinatlas.org/download
#> 13        3.16     NA        TSV https://www.proteinatlas.org/download
#> 14        3.16     NA        TSV https://www.proteinatlas.org/download
#> 15        3.16     NA        TSV https://www.proteinatlas.org/download
#>    SourceVersion      Species TaxonomyId Coordinate_1_based        DataProvider
#> 1           16.1 Homo sapiens       9606                 NA Human Protein Atlas
#> 2           21.0 Homo sapiens       9606                 NA Human Protein Atlas
#> 3           16.1 Homo sapiens       9606                 NA Human Protein Atlas
#> 4           21.0 Homo sapiens       9606                 NA Human Protein Atlas
#> 5           21.0 Homo sapiens       9606                 NA Human Protein Atlas
#> 6           14.0 Homo sapiens       9606                 NA Human Protein Atlas
#> 7           16.1 Homo sapiens       9606                 NA Human Protein Atlas
#> 8           21.0 Homo sapiens       9606                 NA Human Protein Atlas
#> 9           21.0 Homo sapiens       9606                 NA Human Protein Atlas
#> 10          21.0 Homo sapiens       9606                 NA Human Protein Atlas
#> 11          16.1 Homo sapiens       9606                 NA Human Protein Atlas
#> 12          21.0 Homo sapiens       9606                 NA Human Protein Atlas
#> 13          21.0 Homo sapiens       9606                 NA Human Protein Atlas
#> 14          21.0 Homo sapiens       9606                 NA Human Protein Atlas
#> 15          21.0 Homo sapiens       9606                 NA Human Protein Atlas
#>                                    Maintainer RDataClass DispatchClass
#> 1  Laurent Gatto <laurent.gatto@uclouvain.be> data.frame           Rda
#> 2  Laurent Gatto <laurent.gatto@uclouvain.be> data.frame           Rda
#> 3  Laurent Gatto <laurent.gatto@uclouvain.be> data.frame           Rda
#> 4  Laurent Gatto <laurent.gatto@uclouvain.be> data.frame           Rda
#> 5  Laurent Gatto <laurent.gatto@uclouvain.be> data.frame           Rda
#> 6  Laurent Gatto <laurent.gatto@uclouvain.be> data.frame           Rda
#> 7  Laurent Gatto <laurent.gatto@uclouvain.be> data.frame           Rda
#> 8  Laurent Gatto <laurent.gatto@uclouvain.be> data.frame           Rda
#> 9  Laurent Gatto <laurent.gatto@uclouvain.be> data.frame           Rda
#> 10 Laurent Gatto <laurent.gatto@uclouvain.be> data.frame           Rda
#> 11 Laurent Gatto <laurent.gatto@uclouvain.be> data.frame           Rda
#> 12 Laurent Gatto <laurent.gatto@uclouvain.be> data.frame           Rda
#> 13 Laurent Gatto <laurent.gatto@uclouvain.be> data.frame           Rda
#> 14 Laurent Gatto <laurent.gatto@uclouvain.be> data.frame           Rda
#> 15 Laurent Gatto <laurent.gatto@uclouvain.be> data.frame           Rda
#>                              RDataPath
#> 1          hpar/data/hpaCancer16.1.rda
#> 2              hpar/data/hpaCancer.rda
#> 3    hpar/data/hpaNormalTissue16.1.rda
#> 4        hpar/data/hpaNormalTissue.rda
#> 5           hpar/data/hpaSecretome.rda
#> 6    hpar/data/hpaSubcellularLoc14.rda
#> 7  hpar/data/hpaSubcellularLoc16.1.rda
#> 8      hpar/data/hpaSubcellularLoc.rda
#> 9     hpar/data/rnaConsensusTissue.rda
#> 10       hpar/data/rnaFantomTissue.rda
#> 11   hpar/data/rnaGeneCellLine16.1.rda
#> 12       hpar/data/rnaGeneCellLine.rda
#> 13     hpar/data/rnaGeneTissue21.0.rda
#> 14         hpar/data/rnaGtexTissue.rda
#> 15          hpar/data/rnaHpaTissue.rda
#>                                                                                                             Tags
#> 1  ExperimentHub:ExperimentData:ReproducibleResearch:RepositoryData:CancerData:Homo_sapiens_Data:Proteome:Tissue
#> 2  ExperimentHub:ExperimentData:ReproducibleResearch:RepositoryData:CancerData:Homo_sapiens_Data:Proteome:Tissue
#> 3  ExperimentHub:ExperimentData:ReproducibleResearch:RepositoryData:CancerData:Homo_sapiens_Data:Proteome:Tissue
#> 4  ExperimentHub:ExperimentData:ReproducibleResearch:RepositoryData:CancerData:Homo_sapiens_Data:Proteome:Tissue
#> 5  ExperimentHub:ExperimentData:ReproducibleResearch:RepositoryData:CancerData:Homo_sapiens_Data:Proteome:Tissue
#> 6  ExperimentHub:ExperimentData:ReproducibleResearch:RepositoryData:CancerData:Homo_sapiens_Data:Proteome:Tissue
#> 7  ExperimentHub:ExperimentData:ReproducibleResearch:RepositoryData:CancerData:Homo_sapiens_Data:Proteome:Tissue
#> 8  ExperimentHub:ExperimentData:ReproducibleResearch:RepositoryData:CancerData:Homo_sapiens_Data:Proteome:Tissue
#> 9  ExperimentHub:ExperimentData:ReproducibleResearch:RepositoryData:CancerData:Homo_sapiens_Data:Proteome:Tissue
#> 10 ExperimentHub:ExperimentData:ReproducibleResearch:RepositoryData:CancerData:Homo_sapiens_Data:Proteome:Tissue
#> 11 ExperimentHub:ExperimentData:ReproducibleResearch:RepositoryData:CancerData:Homo_sapiens_Data:Proteome:Tissue
#> 12 ExperimentHub:ExperimentData:ReproducibleResearch:RepositoryData:CancerData:Homo_sapiens_Data:Proteome:Tissue
#> 13 ExperimentHub:ExperimentData:ReproducibleResearch:RepositoryData:CancerData:Homo_sapiens_Data:Proteome:Tissue
#> 14 ExperimentHub:ExperimentData:ReproducibleResearch:RepositoryData:CancerData:Homo_sapiens_Data:Proteome:Tissue
#> 15 ExperimentHub:ExperimentData:ReproducibleResearch:RepositoryData:CancerData:Homo_sapiens_Data:Proteome:Tissue

getHpaVersion()
#> version 
#>  "21.1" 
getHpaDate()
#>         date 
#> "2022.05.31" 
getHpaEnsembl()
#>  ensembl 
#> "103.38"