This package provides a simple interface to the Human Protein Atlas. From the Human Protein Atlas Project page: The Swedish Human Protein Atlas project, funded by the Knut and Alice Wallenberg Foundation, has been set up to allow for a systematic exploration of the human proteome using Antibody-Based Proteomics. This is accomplished by combining high-throughput generation of affinity-purified antibodies with protein profiling in a multitude of tissues and cells assembled in tissue microarrays. Confocal microscopy analysis using human cell lines is performed for more detailed protein localization. The program hosts the Human Protein Atlas portal with expression profiles of human proteins in tissues and cells.

Details

Several flat files are distributed by the HPA project and available within the package as data.frames. The description below is taken from the HPA site:

hpaNormalTissue

Normal tissue data: Expression profiles for proteins in human tissues based on immunohistochemisty using tissue micro arrays. The comma-separated file includes Ensembl gene identifier ("Gene"), tissue name ("Tissue"), annotated cell type ("Cell type"), expression value ("Level"), the type of annotation (annotated protein expression (APE), based on more than one antibody, or staining, based on one antibody only) ("Expression type"), and the reliability or validation of the expression value ("Reliability").

hpaCancer

Cancer tumor data: Staining profiles for proteins in human tumor tissue based on immunohistochemisty using tissue micro arrays. The comma-separated file includes Ensembl gene identifier ("Gene"), tumor name ("Tumor"), staining value ("Level"), the number of patients that stain for this staining value ("Count patients"), the total amount of patients for this tumor type ("Total patients") and the type of annotation staining ("Expression type").

rnaGeneTissue

RNA gene data: RNA levels in 45 cell lines and 32 tissues based on RNA-seq. The comma-separated file includes Ensembl gene identifier ("Gene"), analysed sample ("Sample"), fragments per kilobase of transcript per million fragments mapped ("Value" and "Unit"), and abundance class ("Abundance").

rnaGeneCellLine

RNA gene data: RNA levels in 45 cell lines and 32 tissues based on RNA-seq. The comma-separated file includes Ensembl gene identifier ("Gene"), analysed sample ("Sample"), fragments per kilobase of transcript per million fragments mapped ("Value" and "Unit"), and abundance class ("Abundance").

hpaSubcellularLoc

Subcellular location data: Subcellular localization of proteins based on immunofluorescently stained cells. The comma-separated file includes Ensembl gene identifier ("Gene"), main subcellular location of the protein ("Main location"), other locations ("Other location"), the type of annotation (annotated protein expression (APE), based on more than one antibody, or staining, based on one antibody only) ("Expression type"), and the reliability or validation of the expression value ("Reliability").

hpaSubcellularLoc14

Same as above, for version 14.

Detailed description for gene entries and images and not included in the package but can be accessed from within the R environment through a web browser while on-line.

The full data sets can be individually loaded using the data function (see example below). Data about individual genes of interest can retrived with the getHpa function.

HPA data usage policy: The use of data and images from this site in publications and presentations is permitted provided that the following conditions are met:

The publication and/or presentation are solely for informational and non-commercial purposes.

The source of the data and/or image is referred to this site (www.proteinatlas.org) and/or one or more of our publications are cited.

See also

getHpaDate for release information. Gene-specific information should be accessed using the getHpa function.

The package vignette can be accessed with vignette("hpar").

References

See the Human Protein Atlas Project page http://www.proteinatlas.org/ and http://www.proteinatlas.org/about/download for more details and documentation.

Uhlen et al (2010). Towards a knowledge-based Human Protein Atlas. Nat Biotechnol. 28(12):1248-50.

Berglund et al (2008). A gene-centric Human Protein Atlas for expression profiles based on antibodies. Mol Cell Proteomics. 7(10):2019-27.

Uhlen et al (2005). A human protein atlas for normal and cancer tissues based on antibody proteomics. Mol Cell Proteomics. 4(12):1920-1932.

Ponten et al (2008). The Human Protein Atlas - a tool for pathology. J Pathology. 216(4):387-93.

Examples

allHparData()

#> [1] "hpaNormalTissue"     "rnaGeneCellLine"     "rnaGeneTissue"      
#> [4] "hpaSubcellularLoc"   "hpaSubcellularLoc14" "hpaCancer"          

data(hpaNormalTissue)
head(hpaNormalTissue)

#>              Gene Gene.name        Tissue           Cell.type        Level
#> 1 ENSG00000000003    TSPAN6 adrenal gland     glandular cells Not detected
#> 2 ENSG00000000003    TSPAN6      appendix     glandular cells       Medium
#> 3 ENSG00000000003    TSPAN6      appendix     lymphoid tissue Not detected
#> 4 ENSG00000000003    TSPAN6   bone marrow hematopoietic cells Not detected
#> 5 ENSG00000000003    TSPAN6        breast          adipocytes Not detected
#> 6 ENSG00000000003    TSPAN6        breast     glandular cells         High
#>   Expression.type Reliability
#> 1             APE   Uncertain
#> 2             APE   Uncertain
#> 3             APE   Uncertain
#> 4             APE   Uncertain
#> 5             APE   Uncertain
#> 6             APE   Uncertain

data(rnaGeneCellLine)
head(rnaGeneCellLine)

#>              Gene Gene.name  Sample Value Unit Abundance
#> 1 ENSG00000000003    TSPAN6   A-431  17.5 FPKM       Low
#> 2 ENSG00000000003    TSPAN6    A549  27.2 FPKM    Medium
#> 3 ENSG00000000003    TSPAN6  AN3-CA  32.9 FPKM    Medium
#> 4 ENSG00000000003    TSPAN6    BEWO  26.0 FPKM    Medium
#> 5 ENSG00000000003    TSPAN6  CACO-2  56.2 FPKM      High
#> 6 ENSG00000000003    TSPAN6 CAPAN-2  30.0 FPKM    Medium

data(rnaGeneTissue)
head(rnaGeneTissue)

#>              Gene Gene.name          Sample Value Unit    Abundance
#> 1 ENSG00000000003    TSPAN6  adipose tissue  20.5 FPKM       Medium
#> 2 ENSG00000000003    TSPAN6   adrenal gland  17.6 FPKM       Medium
#> 3 ENSG00000000003    TSPAN6        appendix   6.4 FPKM          Low
#> 4 ENSG00000000003    TSPAN6     bone marrow   0.4 FPKM Not detected
#> 5 ENSG00000000003    TSPAN6 cerebral cortex   9.3 FPKM          Low
#> 6 ENSG00000000003    TSPAN6           colon  41.2 FPKM       Medium

data(hpaSubcellularLoc)
head(hpaSubcellularLoc)

#>              Gene Gene.name                         Main.location
#> 1 ENSG00000000003    TSPAN6                             Cytoplasm
#> 2 ENSG00000000457     SCYL3           Cytoskeleton (Microtubules)
#> 3 ENSG00000000460  C1orf112 Nucleus but not nucleoli;Mitochondria
#> 4 ENSG00000000938       FGR                       Plasma membrane
#> 5 ENSG00000001036     FUCA2 Nucleus but not nucleoli;Mitochondria
#> 6 ENSG00000001084      GCLC            Nucleus;Nucleoli;Cytoplasm
#>                             Other.location Expression.type Reliability
#> 1                                                      APE   Uncertain
#> 2 Nucleus but not nucleoli;Golgi apparatus             APE   Uncertain
#> 3                                                      APE   Uncertain
#> 4                                Aggresome             APE   Uncertain
#> 5                                                      APE   Uncertain
#> 6                                                      APE  Supportive
#>                Main.location.GO.id  Other.location.GO.id
#> 1                       GO:0005737                      
#> 2                       GO:0015630 GO:0005654;GO:0005794
#> 3            GO:0005654;GO:0005739                      
#> 4                       GO:0005886            GO:0016235
#> 5            GO:0005654;GO:0005739                      
#> 6 GO:0005634;GO:0005730;GO:0005737                      

data(hpaSubcellularLoc14)
head(hpaSubcellularLoc14)

#>              Gene                         Main.location
#> 1 ENSG00000000003                             Cytoplasm
#> 2 ENSG00000000457           Cytoskeleton (Microtubules)
#> 3 ENSG00000000460 Nucleus but not nucleoli;Mitochondria
#> 4 ENSG00000001036 Nucleus but not nucleoli;Mitochondria
#> 5 ENSG00000001084                      Nucleus;Nucleoli
#> 6 ENSG00000001460                               Nucleus
#>                             Other.location Expression.type Reliability
#> 1                                                      APE   Uncertain
#> 2 Nucleus but not nucleoli;Golgi apparatus             APE   Uncertain
#> 3                                                      APE   Uncertain
#> 4                                                      APE   Uncertain
#> 5                                Cytoplasm             APE  Supportive
#> 6                         Nuclear membrane             APE  Supportive

data(hpaCancer)
head(hpaCancer)

#>              Gene Gene.name         Tumor        Level Count.patients
#> 1 ENSG00000000003    TSPAN6 breast cancer         High              1
#> 2 ENSG00000000003    TSPAN6 breast cancer       Medium              7
#> 3 ENSG00000000003    TSPAN6 breast cancer          Low              2
#> 4 ENSG00000000003    TSPAN6 breast cancer Not detected              2
#> 5 ENSG00000000003    TSPAN6     carcinoid         High              0
#> 6 ENSG00000000003    TSPAN6     carcinoid       Medium              1
#>   Total.patients Expression.type
#> 1             12        Staining
#> 2             12        Staining
#> 3             12        Staining
#> 4             12        Staining
#> 5              4        Staining
#> 6              4        Staining