HPA gene query

Queries one if the HPA data sets with the id Ensembl gene identifier. The data set to be used is defined by the hpadata argument.

getHpa(id, hpadata = NULL, type = c("data", "details"))

Arguments

id
A Ensembl gene identifier.

hpadata
A character with the data set to query. One of available hpar datasets, available by calling allHparData (or any unambiguous prefix), or NULL (default). In the latter case, the default option is used. See setHparOptions and getHparOptions for more details.

type
A character defining what type data to return. One of data (default, for the data as data.frame) or details (to open the id's HPA webpage).

Value

A data.frame with the information corresponding to the respective id genes. If type is details, then the dataframe is returned invisibly and a web page is opened with browseURL.

Author

Laurent Gatto

Examples

id <- "ENSG00000000003"
## Define 'hpadata' data manually
getHpa(id, hpadata = "hpaSubcellularLoc")

#>              Gene Gene.name Reliability          Main.location
#> 1 ENSG00000000003    TSPAN6    Approved Cell Junctions;Cytosol
#>         Additional.location Extracellular.location Enhanced Supported
#> 1 Nucleoli fibrillar center                                          
#>                                           Approved Uncertain
#> 1 Cell Junctions;Cytosol;Nucleoli fibrillar center          
#>   Single.cell.variation.intensity Single.cell.variation.spatial
#> 1                         Cytosol                              
#>   Cell.cycle.dependency
#> 1                      
#>                                                                                     GO.id
#> 1 Cell Junctions (GO:0030054);Cytosol (GO:0005829);Nucleoli fibrillar center (GO:0001650)

head(getHpa(id, hpadata = "hpaNormalTissue"), )

#>              Gene Gene.name         Tissue           Cell.type        Level
#> 1 ENSG00000000003    TSPAN6 adipose tissue          adipocytes Not detected
#> 2 ENSG00000000003    TSPAN6  adrenal gland     glandular cells Not detected
#> 3 ENSG00000000003    TSPAN6       appendix     glandular cells       Medium
#> 4 ENSG00000000003    TSPAN6       appendix     lymphoid tissue Not detected
#> 5 ENSG00000000003    TSPAN6    bone marrow hematopoietic cells Not detected
#> 6 ENSG00000000003    TSPAN6         breast          adipocytes Not detected
#>   Reliability
#> 1    Approved
#> 2    Approved
#> 3    Approved
#> 4    Approved
#> 5    Approved
#> 6    Approved

head(getHpa(id, hpadata = "rnaGeneTissue"))

#>              Gene Gene.name          Tissue  TPM pTPM
#> 1 ENSG00000000003    TSPAN6  adipose tissue 31.5 37.7
#> 2 ENSG00000000003    TSPAN6   adrenal gland 26.4 32.7
#> 3 ENSG00000000003    TSPAN6        appendix  9.2 14.5
#> 4 ENSG00000000003    TSPAN6     bone marrow  0.7  0.9
#> 5 ENSG00000000003    TSPAN6          breast 53.3 68.1
#> 6 ENSG00000000003    TSPAN6 cerebral cortex 18.5 22.8

head(getHpa(id, hpadata = "rnaGeneCellLine"))

#>              Gene Gene.name Cell.line   TPM  pTPM   NX
#> 1 ENSG00000000003    TSPAN6     A-431  27.8  33.9  7.8
#> 2 ENSG00000000003    TSPAN6      A549  37.6  45.5 10.5
#> 3 ENSG00000000003    TSPAN6      AF22 108.1 134.5 28.3
#> 4 ENSG00000000003    TSPAN6    AN3-CA  51.8  64.4 14.3
#> 5 ENSG00000000003    TSPAN6  ASC diff  32.3  37.4 12.4
#> 6 ENSG00000000003    TSPAN6 ASC TERT1  17.7  20.8  6.7

head(getHpa(id, hpadata = "hpaCancer"))

#>              Gene Gene.name             Cancer High Medium Low Not.detected
#> 1 ENSG00000000003    TSPAN6      breast cancer    1      7   2            2
#> 2 ENSG00000000003    TSPAN6          carcinoid    0      1   1            2
#> 3 ENSG00000000003    TSPAN6    cervical cancer   11      1   0            0
#> 4 ENSG00000000003    TSPAN6  colorectal cancer    0      6   2            2
#> 5 ENSG00000000003    TSPAN6 endometrial cancer   10      2   0            0
#> 6 ENSG00000000003    TSPAN6             glioma    0      0   0           11
#>   prognostic.favourable unprognostic.favourable prognostic.unfavourable
#> 1                    NA                0.077120                      NA
#> 2                    NA                      NA                      NA
#> 3                    NA                0.089670                      NA
#> 4                    NA                0.035620                      NA
#> 5                    NA                      NA                      NA
#> 6                    NA                0.002707                      NA
#>   unprognostic.unfavourable
#> 1                        NA
#> 2                        NA
#> 3                        NA
#> 4                        NA
#> 5                    0.2567
#> 6                        NA

## Sets default to "SubcellularLoc"
setHparOptions(hpadata = "hpaSubcellularLoc")
getHpa(id)       ## now uses "hpaSubcellularLoc"

#>              Gene Gene.name Reliability          Main.location
#> 1 ENSG00000000003    TSPAN6    Approved Cell Junctions;Cytosol
#>         Additional.location Extracellular.location Enhanced Supported
#> 1 Nucleoli fibrillar center                                          
#>                                           Approved Uncertain
#> 1 Cell Junctions;Cytosol;Nucleoli fibrillar center          
#>   Single.cell.variation.intensity Single.cell.variation.spatial
#> 1                         Cytosol                              
#>   Cell.cycle.dependency
#> 1                      
#>                                                                                     GO.id
#> 1 Cell Junctions (GO:0030054);Cytosol (GO:0005829);Nucleoli fibrillar center (GO:0001650)

setHparOptions() ## reset to "hpaNormalTissue"
head(getHpa(id))

#>              Gene Gene.name         Tissue           Cell.type        Level
#> 1 ENSG00000000003    TSPAN6 adipose tissue          adipocytes Not detected
#> 2 ENSG00000000003    TSPAN6  adrenal gland     glandular cells Not detected
#> 3 ENSG00000000003    TSPAN6       appendix     glandular cells       Medium
#> 4 ENSG00000000003    TSPAN6       appendix     lymphoid tissue Not detected
#> 5 ENSG00000000003    TSPAN6    bone marrow hematopoietic cells Not detected
#> 6 ENSG00000000003    TSPAN6         breast          adipocytes Not detected
#>   Reliability
#> 1    Approved
#> 2    Approved
#> 3    Approved
#> 4    Approved
#> 5    Approved
#> 6    Approved

## multiple ids
getHpa(id = c("ENSG00000000003", "ENSG00000000005"),
       hpadata = "hpaSubcellularLoc")

#>              Gene Gene.name Reliability          Main.location
#> 1 ENSG00000000003    TSPAN6    Approved Cell Junctions;Cytosol
#>         Additional.location Extracellular.location Enhanced Supported
#> 1 Nucleoli fibrillar center                                          
#>                                           Approved Uncertain
#> 1 Cell Junctions;Cytosol;Nucleoli fibrillar center          
#>   Single.cell.variation.intensity Single.cell.variation.spatial
#> 1                         Cytosol                              
#>   Cell.cycle.dependency
#> 1                      
#>                                                                                     GO.id
#> 1 Cell Junctions (GO:0030054);Cytosol (GO:0005829);Nucleoli fibrillar center (GO:0001650)

if (FALSE) {
## opens a browser with http://www.proteinatlas.org/ENSG00000163435
getHpa("ENSG00000163435", type = "details") 
}