Human Protein Atlas in R

This package provides a simple interface to the Human Protein Atlas. From the Human Protein Atlas Project page: The Swedish Human Protein Atlas project, funded by the Knut and Alice Wallenberg Foundation, has been set up to allow for a systematic exploration of the human proteome using Antibody-Based Proteomics. This is accomplished by combining high-throughput generation of affinity-purified antibodies with protein profiling in a multitude of tissues and cells assembled in tissue microarrays. Confocal microscopy analysis using human cell lines is performed for more detailed protein localization. The program hosts the Human Protein Atlas portal with expression profiles of human proteins in tissues and cells.

Details

Several flat files are distributed by the HPA project and available within the package as data.frames. The description below is taken from the HPA site:

hpaNormalTissue

Normal tissue data: Expression profiles for proteins in human tissues based on immunohistochemisty using tissue micro arrays. The comma-separated file includes Ensembl gene identifier ("Gene"), tissue name ("Tissue"), annotated cell type ("Cell type"), expression value ("Level"), the type of annotation (annotated protein expression (APE), based on more than one antibody, or staining, based on one antibody only) ("Expression type"), and the reliability or validation of the expression value ("Reliability").

hpaCancer

Cancer tumor data: Staining profiles for proteins in human tumor tissue based on immunohistochemisty using tissue micro arrays. The comma-separated file includes Ensembl gene identifier ("Gene"), tumor name ("Tumor"), staining value ("Level"), the number of patients that stain for this staining value ("Count patients"), the total amount of patients for this tumor type ("Total patients") and the type of annotation staining ("Expression type").

rnaGeneTissue

RNA gene data: RNA levels in 45 cell lines and 32 tissues based on RNA-seq. The comma-separated file includes Ensembl gene identifier ("Gene"), analysed sample ("Sample"), fragments per kilobase of transcript per million fragments mapped ("Value" and "Unit"), and abundance class ("Abundance").

rnaGeneCellLine

RNA gene data: RNA levels in 45 cell lines and 32 tissues based on RNA-seq. The comma-separated file includes Ensembl gene identifier ("Gene"), analysed sample ("Sample"), fragments per kilobase of transcript per million fragments mapped ("Value" and "Unit"), and abundance class ("Abundance").

hpaSubcellularLoc

Subcellular location data: Subcellular localization of proteins based on immunofluorescently stained cells. The comma-separated file includes Ensembl gene identifier ("Gene"), main subcellular location of the protein ("Main location"), other locations ("Other location"), the type of annotation (annotated protein expression (APE), based on more than one antibody, or staining, based on one antibody only) ("Expression type"), and the reliability or validation of the expression value ("Reliability").

hpaSubcellularLoc14

Same as above, for version 14.

Detailed description for gene entries and images and not included in the package but can be accessed from within the R environment through a web browser while on-line.

The full data sets can be individually loaded using the data function (see example below). Data about individual genes of interest can retrived with the getHpa function.

HPA data usage policy: The use of data and images from this site in publications and presentations is permitted provided that the following conditions are met:

The publication and/or presentation are solely for informational and non-commercial purposes.

The source of the data and/or image is referred to this site (www.proteinatlas.org) and/or one or more of our publications are cited.

See also

References

See the Human Protein Atlas Project page http://www.proteinatlas.org/ and http://www.proteinatlas.org/about/download for more details and documentation.

Uhlen et al (2010). Towards a knowledge-based Human Protein Atlas. Nat Biotechnol. 28(12):1248-50.

Berglund et al (2008). A gene-centric Human Protein Atlas for expression profiles based on antibodies. Mol Cell Proteomics. 7(10):2019-27.

Uhlen et al (2005). A human protein atlas for normal and cancer tissues based on antibody proteomics. Mol Cell Proteomics. 4(12):1920-1932.

Ponten et al (2008). The Human Protein Atlas - a tool for pathology. J Pathology. 216(4):387-93.

Examples

allHparData()

#> [1] "hpaNormalTissue"     "rnaGeneCellLine"     "rnaGeneTissue"      
#> [4] "hpaSubcellularLoc"   "hpaSubcellularLoc14" "hpaCancer"          

data(hpaNormalTissue)
head(hpaNormalTissue)

#>              Gene Gene.name        Tissue           Cell.type        Level
#> 1 ENSG00000000003    TSPAN6 adrenal gland     glandular cells Not detected
#> 2 ENSG00000000003    TSPAN6      appendix     glandular cells       Medium
#> 3 ENSG00000000003    TSPAN6      appendix     lymphoid tissue Not detected
#> 4 ENSG00000000003    TSPAN6   bone marrow hematopoietic cells Not detected
#> 5 ENSG00000000003    TSPAN6        breast          adipocytes Not detected
#> 6 ENSG00000000003    TSPAN6        breast     glandular cells         High
#>   Reliability
#> 1    Approved
#> 2    Approved
#> 3    Approved
#> 4    Approved
#> 5    Approved
#> 6    Approved

data(rnaGeneCellLine)
head(rnaGeneCellLine)

#>              Gene Gene.name    Sample Value Unit
#> 1 ENSG00000000003    TSPAN6     A-431  27.8  TPM
#> 2 ENSG00000000003    TSPAN6      A549  37.6  TPM
#> 3 ENSG00000000003    TSPAN6      AF22 108.1  TPM
#> 4 ENSG00000000003    TSPAN6    AN3-CA  51.8  TPM
#> 5 ENSG00000000003    TSPAN6  ASC diff  32.3  TPM
#> 6 ENSG00000000003    TSPAN6 ASC TERT1  17.7  TPM

data(rnaGeneTissue)
head(rnaGeneTissue)

#>              Gene Gene.name          Sample Value Unit
#> 1 ENSG00000000003    TSPAN6  adipose tissue  31.5  TPM
#> 2 ENSG00000000003    TSPAN6   adrenal gland  26.4  TPM
#> 3 ENSG00000000003    TSPAN6        appendix   9.2  TPM
#> 4 ENSG00000000003    TSPAN6     bone marrow   0.7  TPM
#> 5 ENSG00000000003    TSPAN6          breast  53.4  TPM
#> 6 ENSG00000000003    TSPAN6 cerebral cortex  18.5  TPM

data(hpaSubcellularLoc)
head(hpaSubcellularLoc)

#>              Gene Gene.name Reliability Enhanced Supported
#> 1 ENSG00000000003    TSPAN6    Approved                   
#> 2 ENSG00000000457     SCYL3   Uncertain                   
#> 3 ENSG00000000460  C1orf112    Approved                   
#> 4 ENSG00000000938       FGR    Approved                   
#> 5 ENSG00000000971       CFH    Approved                   
#> 6 ENSG00000001084      GCLC    Approved                   
#>                    Approved                   Uncertain
#> 1                   Cytosol                            
#> 2                           Microtubules;Nuclear bodies
#> 3              Mitochondria                            
#> 4 Aggresome;Plasma membrane                            
#> 5                  Vesicles                            
#> 6  Cytosol;Nucleoli;Nucleus                            
#>   Single.cell.variation.intensity Single.cell.variation.spatial
#> 1                                                              
#> 2                                                              
#> 3                                                              
#> 4                                                              
#> 5                                                              
#> 6                                                              
#>   Cell.cycle.dependency
#> 1                      
#> 2                      
#> 3                      
#> 4                      
#> 5                      
#> 6                      
#>                                                             GO.id
#> 1                                            Cytosol (GO:0005829)
#> 2           Microtubules (GO:0015630);Nuclear bodies (GO:0016604)
#> 3                                       Mitochondria (GO:0005739)
#> 4             Aggresome (GO:0016235);Plasma membrane (GO:0005886)
#> 5                                           Vesicles (GO:0043231)
#> 6 Cytosol (GO:0005829);Nucleoli (GO:0005730);Nucleus (GO:0005634)

data(hpaSubcellularLoc14)
head(hpaSubcellularLoc14)

#>              Gene                         Main.location
#> 1 ENSG00000000003                             Cytoplasm
#> 2 ENSG00000000457           Cytoskeleton (Microtubules)
#> 3 ENSG00000000460 Nucleus but not nucleoli;Mitochondria
#> 4 ENSG00000001036 Nucleus but not nucleoli;Mitochondria
#> 5 ENSG00000001084                      Nucleus;Nucleoli
#> 6 ENSG00000001460                               Nucleus
#>                             Other.location Expression.type Reliability
#> 1                                                      APE   Uncertain
#> 2 Nucleus but not nucleoli;Golgi apparatus             APE   Uncertain
#> 3                                                      APE   Uncertain
#> 4                                                      APE   Uncertain
#> 5                                Cytoplasm             APE  Supportive
#> 6                         Nuclear membrane             APE  Supportive

data(hpaCancer)
head(hpaCancer)

#>              Gene Gene.name             Cancer High Medium Low Not.detected
#> 1 ENSG00000000003    TSPAN6      breast cancer    1      7   2            2
#> 2 ENSG00000000003    TSPAN6          carcinoid    0      1   1            2
#> 3 ENSG00000000003    TSPAN6    cervical cancer   11      1   0            0
#> 4 ENSG00000000003    TSPAN6  colorectal cancer    0      6   2            2
#> 5 ENSG00000000003    TSPAN6 endometrial cancer   10      2   0            0
#> 6 ENSG00000000003    TSPAN6             glioma    0      0   0           11
#>   prognostic.favourable unprognostic.favourable prognostic.unfavourable
#> 1                    NA                0.077120                      NA
#> 2                    NA                      NA                      NA
#> 3                    NA                0.089670                      NA
#> 4                    NA                0.035620                      NA
#> 5                    NA                      NA                      NA
#> 6                    NA                0.003617                      NA
#>   unprognostic.unfavourable
#> 1                        NA
#> 2                        NA
#> 3                        NA
#> 4                        NA
#> 5                    0.2567
#> 6                        NA